Revert "separate indexMethod"
