+++ index.js
@@ -4,39 +4,91 @@ var mysql = require('mysql');
 var http = require('http');
 var fs = require('fs');
 var url = require('url');
-var db = require('./public/js/db.js');
+
 
 var app = express();
-var template = require('./public/js/template.js');
+
+
+
 
 app.use(express.static(__dirname + '/public'));
-// // html 파일 읽기위해 사용함
+// html 파일 읽기위해 사용함
 app.set('views', __dirname + '/views');
 app.engine('html', engine.mustache);
 app.set('view engine', 'html');
-// 나는 이거 왜 필요한지 모르겠다.. 선장님이 좀 알려주셈/ 없어도 실행 잘 되긴함.
+// 여기까지
 
-var app= http.createServer('/',function(req,res){
+function templateHTML(title,list,feature) {
+    return `
+    <!DOCTYPE html>
+    <html>
+    <head>
+        <meta charset="utf-8" />
+        <link href="css/main.css?ver=1" type="text/css" rel="stylesheet" />
+    </head>
+
+    <body>
+        <h1><a href="">This is ${title}</a></h1>
+        ${list}
+        <a href="?id=login">LOGIN</a>
+        <a href="?id=contact">CONTACT</a>
+        ${feature}
+
+    </body>
+    </html>
+
+    `;
+}
+
+function templateList(filelist) {
+    var list = '<ul>';
+    var i = 0;
+    while(i < filelist.length){
+        var title = filelist[i].split('.')[0];
+        list = list + `<li><a href="/?id=${title}">${title}</a></li>`;
+        i = i + 1;
+    }
+    list = list+'</ul>';
+    return list;
+}
+
+var connection = mysql.createConnection({
+	host: 'localhost',
+	user: 'root',
+
+	post: 3306,
+
+	password: 'oracle11',
+	database: 'my_db'
+});
+
+app.get('/',function(req,res){
     var _url = req.url;
     var queryData = url.parse(_url, true).query;
-        if(queryData.id === undefined){
-          fs.readdir('./views', function(error, filelist){
-          var title = 'main';
-          var list = template.list(filelist);
-          var html = template.HTML(title,list,`<h2>${title}</h2>`);
-          res.writeHead(200);
-          res.end(html);
-          })
+    fs.readdir('views', function(error, filelist){
+        var title = 'main';
+        var template;
+        var list;
+        if(req.query.id === undefined || req.query.id === 'main'){
+            list = templateList(filelist)
+            template = templateHTML(title,list,`<h2>${title}</h2>`);
+            res.end(template);
         }
         else{
-          fs.readdir('./views', function(error, filelist){
             fs.readFile(`views/${queryData.id}.html`, 'utf8', function(err, description){
                 title = queryData.id;
-                list = template.list(filelist);
-                template = template.HTML(title,list,description);
+                list = templateList(filelist)
+                template = templateHTML(title,list,description);
                 res.end(template);
             });
-            });
         }
+    })
+
+
 });
-app.listen(3000);
+
+
+
+app.listen(3000,function () {
+    console.log('Server start, port : 3000');
+});
