+++ views/main.html
@@ -1,4 +1,4 @@
-<!-- <!DOCTYPE html>
+<!DOCTYPE html>
 <html>
 <head>
 	<meta charset="utf-8" />
@@ -18,4 +18,4 @@
 			<a href="?id=login">LOGIN</a>
 			<a href="?id=contact">CONTACT</a>
 </body>
-</html> -->
+</html>
